fixed bug in beam search.  Tested.
The bug was caused by floating point error resulting from the use of softmax
instead of log-softmax to choose the next word in the beam.  Changing to log
softmax fixed it.

Also, I tested that beam search with width 1 yields the same result as
the old predict methods for EncoderDecoderRNN and SearchRN.  This test
was passed.

After I have done a bit more testing, I intend to delete the old
predict methods, as beam search can completely replace them.